Only generate Redis config URL if not present in env
Contrary to the initial implementation the Redis config URL will not always be updated with Redis environment variables. Instead the Redis config URL can include bash style placeholders ${VAR} that reference those environment variables. If `AUTHENTIK_REDIS__URL` is not set i.e. in an old Authentik setup the Redis config URL is automatically generated and a warning message is issued to make the user aware of the necessary future configuration changes.