I'm conducting a more comprehensive survey of the UI in order to get a more holistic idea of the changes that should be implemented.  Along the way, I'm finding a few small details that annoy me. Here are three.
It goes to "User statistics."  I have changed both the text of the link and the page to read "User
Statistics" (it's a title, it should be capitalized).

Give people warning when you're about to take them out of the system, especially if you're opening a
new tab along the way.

Just a thing I spotted along the way.